% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{accept}
\alias{accept}
\title{Accept a simulated movement}
\usage{
accept(leave_bounds, allow_empties, within, populated)
}
\arguments{
\item{leave_bounds}{boolean, are movements to outside of the boundaries of the are being
simulated valid}

\item{allow_empties}{boolean, are movements to empty patches (i.e. with no dogs) valid}

\item{within}{boolean vector, whether movement falls within bounds}

\item{populated}{boolean vector, whether movement falls within a populated area}
}
\value{
a boolean vector of length \code{within}/\code{populated} corresponding to whether
a movement is accepted as valid
}
\description{
Helper function to determine if a simulated movement is valid.
To do: tests (same length as within and boolean no NAs)
}
\keyword{internal}
\keyword{transmit}
