% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_continuous.R
\name{angle_fun}
\alias{angle_fun}
\title{Helper function for generating angles to move at}
\usage{
angle_fun(n)
}
\description{
Helper function for generating angles to move at
}
\keyword{internal}
\keyword{move}
