% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{binom_detect}
\alias{binom_detect}
\title{Observation function: single binomial probability}
\usage{
binom_detect(I_dt, params = list(detect_prob = 0.9))
}
\arguments{
\item{I_dt}{the line list of cases from \code{simrabid}}

\item{params}{list of parameters}
}
\value{
The reporting function should modify the line list
in place within the simrabid function (adding a \code{detected} column
as a boolean vector).
}
\description{
This function simulates the observation proccess given a
single detection probability. Defaults to 0.9 when passing p
\code{param_defaults}.
Can be passed to the \code{observe_fun} in \code{\link{simrabid}}.
}
