% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_prob.R
\name{cell_weights}
\alias{cell_weights}
\title{Get cell weights based on covariates or a vector of null probabilities}
\usage{
cell_weights(
  covars = list(0),
  params = list(0),
  ncells,
  leave_bounds,
  allow_invalid,
  cells_block,
  cells_out_bounds
)
}
\arguments{
\item{covars}{list of vectors corresponding to covariates for each cell,
one should be the intercept (valued at 1)}

\item{params}{the parameter corresponding to the effect of each covariate,
one should be length 1 and correspond to the intercept}

\item{ncells}{numeric, the number of total cells (ncol * nrow)}

\item{leave_bounds}{boolean, are movements to outside of the boundaries of the are being simulated are valid}

\item{allow_invalid}{boolean, are movements to empty patches (i.e. with no dogs or bodies of water, etc.) valid}

\item{cells_block}{integer vector, the cell ids of grid cells where movements are invalid to,
for use with max_tries
@param cells_out_bounds integer vector, the cell ids of grid cells which are outside the bounds/
not covered by the area being simulated}
}
\value{
a vector of weights of length ncells + 1
}
\description{
This function generates probabilities for each cell given covariates and
parameter estimates in a logistic regression framework.
As a default it generates uniform probabilities
for each cell (0.5, with invalid and out-of-bounds set to zero if leave_bounds
or allow_invalid are false respectively). Covaraites and parameters can be passed,
but these must include a covariate length 1 equal to 1L and a parameter estimate
of length 1 corresponding to the model intercept. The last weight is for those
cell ids that fall outside the range of possible cell ids (i.e. not in 1:ncells).
}
