% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_prob.R
\name{cells_away}
\alias{cells_away}
\title{Generate cell ids for sampling movement to}
\usage{
cells_away(x0, y0, dist_m, res_m, ncols, nrows, x_topl, y_topl, ncells)
}
\arguments{
\item{res_m}{numeric, the grid cell resolution in meters}

\item{ncols}{numeric, the number of columns in the grid}

\item{nrows}{numeric, the number of rows in the grid}

\item{x_topl}{numeric, the top left x coordinate of the grid on which
movement is being simulated}

\item{y_topl}{numeric, the top left y coordinate of the grid on which
movement is being simulated}

\item{ncells}{numeric, the number of total cells (ncol * nrow)}
}
\value{
a list of the candidate cell_ids and the x and y coords of the movement
}
\description{
Generate cell ids for sampling movement to
}
\keyword{internal}
\keyword{move}
