% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{check_sim}
\alias{check_sim}
\title{Quick summary stats of time series}
\usage{
check_sim(names = c("I_dt", "S_mat", "N_mat", "tmax", "days_in_step"))
}
\arguments{
\item{names}{objects to get from \code{simrabid} environment}
}
\value{
time series of monthly cases (total, local, and detected), susceptibles,
population size, and coverage.
}
\description{
Sanity check and also test different settings in the simulation framework.
}
\keyword{summarize}
