% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{dispersal_lognorm}
\alias{dispersal_lognorm}
\title{Dispersal kernel distribution}
\usage{
dispersal_lognorm(n, params)
}
\arguments{
\item{n}{number to draw}

\item{params}{list of parameters}
}
\value{
numeric vector, dispersal distances of length \code{n}
}
\description{
This function draws dispersal distances (in meters)
from a lognormal distribution, pass \code{param_defaults} to the parameter list
for best-fit parameters from Mancy et al.
Can be passed to the \code{dispersal_fun} in \code{\link{simrabid}} (should
set parameter \code{sequential} to FALSE as well in that case).
}
