% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demdata_SD.R
\name{get.demdata}
\alias{get.demdata}
\title{add small prob so that allocates equally if vaccinations happen in place where
no coverage reported in census and can be allocated equally when 'leftovers'}
\usage{
get.demdata(shapefile, res_meters = 1000, pop = pop_data, census = census_data)
}
\description{
add small prob so that allocates equally if vaccinations happen in place where
no coverage reported in census and can be allocated equally when 'leftovers'
}
