% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_prob.R
\name{get_cellweights}
\alias{get_cellweights}
\title{Subset weights of cells to the candidate cell ids}
\usage{
get_cellweights(weights, cell_ids)
}
\arguments{
\item{weights}{numeric vector, weights to give each grid cell for sampling moves
for use with sim_movement_prob; will be length ncells + 1,
see \code{\link{cell_weights}}) for more details.
defaults to NULL}

\item{cell_ids}{integer vector, the cell ids corresponding to each grid cell in which
the infectious individual is starting from}
}
\value{
a vector of weights corresponding to the cell_ids passed through
}
\description{
Subset weights of cells to the candidate cell ids
}
\keyword{internal}
\keyword{move}
