% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init}
\alias{init}
\title{Initialize the simulation}
\usage{
init(
  start_pop,
  start_vacc,
  I_seeds,
  I_dt,
  cell_ids,
  admin_ids = NULL,
  row_ids,
  bins,
  x_coord,
  y_coord,
  params,
  incursion_fun,
  ncells
)
}
\description{
\code{init} starts up the simulation by drawing the starting state values
and seeds infectious cases
}
\keyword{internal}
