% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_continuous.R
\name{move_continuous}
\alias{move_continuous}
\title{Movement in continuous space}
\usage{
move_continuous(
  dist_m,
  angle,
  x0,
  y0,
  x_topl,
  y_topl,
  res_m,
  ncols,
  nrows,
  ncells,
  cells_block,
  cells_out_bounds,
  path
)
}
\arguments{
\item{dist_m}{numeric vector or value, distance to move}

\item{angle}{numeric vector \link{0, 360} or value, angle at which to move at}

\item{x0}{numeric vector or value, the x origin of the infected individual}

\item{y0}{numeric vector or value, the y origin of the infected individual}

\item{x_topl}{numeric, the top left x coordinate of the grid on which
movement is being simulated}

\item{y_topl}{numeric, the top left y coordinate of the grid on which
movement is being simulated}

\item{res_m}{numeric, the grid cell resolution in meters}

\item{ncols}{numeric, the number of columns in the grid}

\item{nrows}{numeric, the number of rows in the grid}

\item{ncells}{numeric, the number of total cells (ncol * nrow)}

\item{cells_block}{integer vector, the cell ids of grid cells where movements are invalid to,
for use with max_tries
@param cells_out_bounds integer vector, the cell ids of grid cells which are outside the bounds/
not covered by the area being simulated}

\item{path}{integer vector or value, if sequential is FALSE then 0, if TRUE then
the path id (i.e. which step in the series of movements) to pass through and assign to exposed}
}
\value{
a list with the resulting x and y coordinates, cell ids, whether
the movement was to a invalid cell, whether the movement fell within the bounds
of the simulation, and the path id for the exposed (can either be length 1 or length of dist_m)
}
\description{
Simulates individual movements
}
\keyword{internal}
