% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_prob.R
\name{movement_prob}
\alias{movement_prob}
\title{Movement based on probability}
\usage{
movement_prob(
  dist_m,
  weights,
  x0,
  y0,
  x_topl,
  y_topl,
  ncells,
  res_m,
  ncols,
  nrows,
  path,
  leave_bounds,
  cells_block,
  cells_out_bounds
)
}
\arguments{
\item{weights}{numeric vector, weights to give each grid cell for sampling moves
for use with sim_movement_prob; will be length ncells + 1,
see \code{\link{cell_weights}}) for more details.
defaults to NULL}

\item{x_topl}{numeric, the top left x coordinate of the grid on which
movement is being simulated}

\item{y_topl}{numeric, the top left y coordinate of the grid on which
movement is being simulated}

\item{ncells}{numeric, the number of total cells (ncol * nrow)}

\item{res_m}{numeric, the grid cell resolution in meters}

\item{ncols}{numeric, the number of columns in the grid}

\item{nrows}{numeric, the number of rows in the grid}

\item{leave_bounds}{boolean, are movements to outside of the boundaries of the are being simulated are valid}

\item{cells_block}{integer vector, the cell ids of grid cells where movements are invalid to,
for use with max_tries
@param cells_out_bounds integer vector, the cell ids of grid cells which are outside the bounds/
not covered by the area being simulated}
}
\value{
list of length 1 with the chosen with sampled cell id,  x and y coordinates, whether
the movement was to a invalid cell, whether the movement fell within the bounds
of the simulation, and the path id for the exposed.
}
\description{
Movement based on probability
}
\keyword{internal}
\keyword{move}
