% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{nbinom_constrained}
\alias{nbinom_constrained}
\title{Secondary case distribution}
\usage{
nbinom_constrained(n, params = list(R0 = 1.2, k = 1, max_secondaries = 100))
}
\arguments{
\item{n}{number to draw}

\item{params}{list of parameters}
}
\value{
integer vector, number of secondary cases, of length \code{n}
}
\description{
This function draws secondary cases from a negative binomial distribution,
but constrained to a maximum (\code{max_secondaries}),
where \code{R0} is the mean (mu in rnbinom) and \code{k} is the dispersion parameter
(size in rnbinom).
Passed to the param \code{secondary_fun} in \code{\link{simrabid}}.
}
