% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{nbinom_constrained_min}
\alias{nbinom_constrained_min}
\title{Secondary case distribution with additional constraint on minimum # of cases}
\usage{
nbinom_constrained_min(
  n,
  params = list(R0 = 1.2, k = 1, max_secodndaries = 100, min_secondaries = 1)
)
}
\arguments{
\item{n}{number to draw}

\item{params}{list of parameters}
}
\value{
integer vector, number of secondary cases, of length \code{n}
}
\description{
This function draws secondary cases from a negative binomial distribution,
but constrained to a maximum (\code{max_secondaries}),
where \code{R0} is the mean (mu in rnbinom) and \code{k} is the dispersion parameter
(size in rnbinom). In addition, if a case is an introduction, secondaries
are additionally constrained to a minimum (\code{min_secondaries}) number of cases.
In general this serves as an example of how to customize functions using
mget and a general template function of function(n, params) {...}.
}
\details{
Passed to the param \code{secondary_fun} in \code{\link{simrabid}}.
}
