% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{return_env}
\alias{return_env}
\title{Return a list of all objects from parent environment}
\usage{
return_env()
}
\value{
a list of all objects from the simrabid environment
}
\description{
This function is an example of a summary function that can be
passed to \code{simrabid}. These functions will be called with no arguments,
so all arguments must have a default or be empty.
}
\keyword{summarize}
