% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_sample}
\alias{safe_sample}
\title{Safer implementation of base sample}
\usage{
safe_sample(x, size, prob = NULL)
}
\description{
Wraps sample function in base R so that if options are length 1 and an integer,
doesn't default to sample.int. See \code{\link{sample}} for arguments. Only
with replace = TRUE (as the function will return the option of length 1 with
length of size).
}
