% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{serial_lognorm}
\alias{serial_lognorm}
\title{Serial interval distribution}
\usage{
serial_lognorm(n, params)
}
\arguments{
\item{n}{number to draw}

\item{params}{list of parameters}
}
\value{
numeric vector, serial intervals of length \code{n}
}
\description{
This function draws serial intervals (in days)
from a lognormal distribution, pass \code{param_defaults} to the parameter list
for best-fit parameters from Mancy et al.
Can be passed to the \code{serial_fun} in \code{\link{simrabid}}.
}
