% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_sim}
\alias{setup_sim}
\title{Set-up simulation}
\usage{
setup_sim(
  tmax,
  rast,
  death_rate_annual,
  birth_rate_annual,
  waning_rate_annual,
  block_fun = block_cells,
  params = list(start_pop),
  step = 52,
  by_admin = FALSE
)
}
\arguments{
\item{tmax}{integer, the maximum number of time steps to run the simulation for}

\item{rast}{raster, from \code{\link{setup_space}} the output raster with each cell allocated
to a location}

\item{death_rate_annual}{annual death rate, either length 1 or length of the admin units}

\item{birth_rate_annual}{annual birth rate, either length 1 or length of the admin units}

\item{waning_rate_annual}{annual waning rate of vaccination immunity,
either length 1 or length of the admin units}

\item{block_fun}{a function that designates the indices to track/block/consider
out-of-bounds (see \code{\link{block_cells}} for an example)}

\item{params}{additional params to pass, must include start_pop, a numeric
vector of the starting population in each grid cell of \code{rast}}

\item{step}{the number of steps to translate rate to probabilities
(i.e. 52 = weekly, 365 = daily)}

\item{by_admin}{logical, should mixing be at the admin (or arbitrary) scale?}
}
\value{
a list of objects needed for the simulation
}
\description{
\code{setup_sim} takes inputs and generates all needed outputs to be passed
to \code{\link{simrabid}}.
}
\keyword{setup}
