% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_space}
\alias{setup_space}
\title{Setup spatial bounds of simulation}
\usage{
setup_space(shapefile, resolution = 1000, id_col, use_fasterize = FALSE)
}
\arguments{
\item{shapefile}{a shapefile (either an sf or sp object) of the area being simulated
should be projected in UTM coordinates}

\item{resolution}{numeric, the resolution in meters}

\item{id_col}{character, the name of the attribute/column to rasterize the data by,
largely for allocating vaccinations}

\item{use_fasterize}{boolean, if TRUE and class of shapefile is sf will use the
fasterize package (if installed)}
}
\value{
raster of simulation bounds rasterized by the id column specify (i.e.
each grid cell allocated to a location)
}
\description{
\code{setup_space} creates the spatial grid of the simulation from a user defined
shapefile and resolution. If vaccination data are aggregated at a larger scale,
pass the id column (i.e. numeric id of the administrative unit) to assign each grid cell to.
Alternatively if vaccination data are aggregated at the grid cell level or vaccination
is absent, then pass NULL to \code{id_col}.
}
\details{
To do:
\itemize{
\item Add examples with sf and sp & different id_col
\item Tests: resolution too large? if no fasterize, if no id col
}
}
\keyword{setup}
