% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce.R
\name{sim_incursions_hardwired}
\alias{sim_incursions_hardwired}
\title{Title}
\usage{
sim_incursions_hardwired(
  nlocs,
  rows_pop,
  row_ids = args$row_ids_empirical,
  tsteps = args$tstep_empirical,
  current_tstep = t
)
}
\arguments{
\item{current_tstep}{}
}
\value{

}
\description{
Title
}
