% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce.R
\name{sim_incursions_pois}
\alias{sim_incursions_pois}
\title{Title}
\usage{
sim_incursions_pois(nlocs, rows_pop, params = list(iota = 4), steps = 4)
}
\arguments{
\item{steps}{}
}
\value{

}
\description{
Title
}
