% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{sim_movement}
\alias{sim_movement}
\title{Simulate individual movements}
\usage{
sim_movement(
  angle,
  distance,
  x0,
  y0,
  x_topl,
  y_topl,
  res_m,
  ncol,
  nrow,
  cells_pop,
  path
)
}
\arguments{
\item{angle}{numeric vector \link{0, 360} or value, angle at which to move at}

\item{distance}{numeric vector or value, distance to move}

\item{x0}{numeric vector or value, the x origin of the infected individual}

\item{y0}{numeric vector or value, the y origin of the infected individual}

\item{x_topl}{numeric, the top left x coordinate of the grid on which
movement is being simulated}

\item{y_topl}{numeric, the top left y coordinate of the grid on which
movement is being simulated}

\item{res_m}{numeric, the grid cell resolution in meters}

\item{ncol}{numeric, the number of columns in the grid}

\item{nrow}{numeric, the number of rows in the grid}

\item{cells_pop}{integer vector, the cell ids of grid cells that are populated
or generally where movements are valid to}

\item{path}{integer vector or value, if sequential is FALSE then 0, if TRUE then
the path id (i.e. which step of the movements) to pass through and assign to exposed}
}
\value{
a list with the resulting x and y coordinates, cell ids, whether
the movement was to a populated cell, whether the movement fell within the bounds
of the simulation, and the path id for the exposed.
}
\description{
Simulates movemement of individuals in continuous space.
}
\details{
This simulates movement and uses the top-left coordinates and 1-based indexing
of raster cell ids to identify the grid cell moved to by an individual.
}
\keyword{internal}
\keyword{transmit}
