% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_continuous.R
\name{sim_movement_continuous}
\alias{sim_movement_continuous}
\title{Simulate individual movements}
\usage{
sim_movement_continuous(
  dist_m,
  angle,
  dispersal_fun,
  x0,
  y0,
  x_topl,
  y_topl,
  res_m,
  ncols,
  nrows,
  ncells,
  cells_block,
  cells_out_bounds,
  path,
  leave_bounds,
  allow_invalid,
  max_tries,
  sequential,
  weights = NULL,
  params
)
}
\arguments{
\item{dist_m}{numeric vector or value, distance to move}

\item{angle}{numeric vector \link{0, 360} or value, angle at which to move at}

\item{x0}{numeric vector or value, the x origin of the infected individual}

\item{y0}{numeric vector or value, the y origin of the infected individual}

\item{x_topl}{numeric, the top left x coordinate of the grid on which
movement is being simulated}

\item{y_topl}{numeric, the top left y coordinate of the grid on which
movement is being simulated}

\item{res_m}{numeric, the grid cell resolution in meters}

\item{ncols}{numeric, the number of columns in the grid}

\item{nrows}{numeric, the number of rows in the grid}

\item{ncells}{numeric, the number of total cells (ncol * nrow)}

\item{cells_block}{integer vector, the cell ids of grid cells where movements are invalid to,
for use with max_tries
@param cells_out_bounds integer vector, the cell ids of grid cells which are outside the bounds/
not covered by the area being simulated}

\item{path}{integer vector or value, if sequential is FALSE then 0, if TRUE then
the path id (i.e. which step in the series of movements) to pass through and assign to exposed}

\item{leave_bounds}{boolean, are movements to outside of the boundaries of the are being simulated are valid}

\item{allow_invalid}{boolean, are movements to empty patches (i.e. with no dogs or bodies of water, etc.) valid}

\item{sequential}{boolean, if TRUE then movements are sequential, if FALSE, then
movements are kernel based}
}
\value{
a list (if sequential) or a data.table (if kernel based) with the resulting x and y coordinates, cell ids, whether
the movement was to a invalid cell, whether the movement fell within the bounds
of the simulation, and the path id for the exposed.
}
\description{
Simulates movemement of individuals in continuous space.
}
\details{
This simulates movement and uses the top-left coordinates and 1-based indexing
of a UTM raster to identify the grid cell moved to by an individual.
}
\keyword{move}
