% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_prob.R
\name{sim_movement_prob}
\alias{sim_movement_prob}
\title{Simulate movement probabilistically}
\usage{
sim_movement_prob(
  dist_m,
  angle = NULL,
  dispersal_fun,
  x0,
  y0,
  x_topl,
  y_topl,
  res_m,
  ncols,
  nrows,
  ncells,
  cells_block,
  cells_out_bounds,
  path,
  leave_bounds,
  allow_invalid,
  max_tries,
  sequential,
  weights = NULL,
  params
)
}
\arguments{
\item{angle}{NULL, dummy to match sim_movement_continuous}

\item{dispersal_fun}{function with a single parameter, n, which will draw distances
for each individual to move in meters}

\item{x_topl}{numeric, the top left x coordinate of the grid on which
movement is being simulated}

\item{y_topl}{numeric, the top left y coordinate of the grid on which
movement is being simulated}

\item{res_m}{numeric, the grid cell resolution in meters}

\item{ncols}{numeric, the number of columns in the grid}

\item{nrows}{numeric, the number of rows in the grid}

\item{ncells}{numeric, the number of total cells (ncol * nrow)}

\item{cells_block}{integer vector, the cell ids of grid cells where movements are invalid to,
for use with max_tries
@param cells_out_bounds integer vector, the cell ids of grid cells which are outside the bounds/
not covered by the area being simulated}

\item{leave_bounds}{boolean, are movements to outside of the boundaries of the are being simulated are valid}

\item{allow_invalid}{boolean, are movements to empty patches (i.e. with no dogs or bodies of water, etc.) valid}

\item{max_tries}{integer, the maximum number of tries to make before accepting
an invalid movement (i.e. transmission event fails due to either leaving the
bounds of the simulation or moving to an uninhabitable grid cell) if either
or both leave_bounds and allow_invalid are FALSE.}

\item{sequential}{boolean, if TRUE then movements are sequential, if FALSE, then
movements are kernel based}

\item{weights}{numeric vector, weights to give each grid cell for sampling moves
for use with sim_movement_prob; will be length ncells + 1,
see \code{\link{cell_weights}}) for more details.
defaults to NULL}
}
\value{
a list (if sequential) or a data.table (if kernel based) with the resulting x and y coordinates, cell ids, whether
the movement was to a invalid cell, whether the movement fell within the bounds
of the simulation, and the path id for the exposed.
}
\description{
Simulate movement probabilistically
}
\keyword{move}
