% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{sim_trans}
\alias{sim_trans}
\title{Simulate transmission outcomes of exposures at individual level}
\usage{
sim_trans(row_id, S, E, I, V, nlocs, track = FALSE)
}
\arguments{
\item{row_id}{numeric vector (length >= 1) the row ids corresponding to the grid cell of the exposure (one value per exposure)}

\item{S, E, I, V}{numeric vector of state variables (i.e. # of individuals in
each class in each grid cell) of length nlocs}

\item{nlocs}{numeric, number of grid cells total}

\item{track}{boolean, whether to explicitly track the outcome of each exposure or
to only track whether successful or not (i.e. with a suscpetible or no). See details.}
}
\value{
a list of two vectors (contact and infected) of same length as \code{row_id}.
contact will be NULL if track = FALSE.
}
\description{
\code{sim_trans} simulates the outcome of each exposure/transmission event
(i.e. whether the exposure was to a susceptible)
}
\details{
This function takes the locations of each exposure/transmission event, tabulates
them, and then allocates them to individuals of each state (using sampling). These
outcomes can be tracked explicitly (i.e. track if S/E/I/V or M if no individuals left to allocate to
in that location) or just whether with a susceptible (S) or not (M) (see \code{track} argument in
\code{\link{simrabid}}.

To Do:
\itemize{
\item Tests: returns should be same length as length of row_id and should be S/E/I/V/M or S/M (no NAs)
\item rewrite in Rcpp? Will it actually speed this up? Maybe by reducing cost of function calls
}
}
\keyword{internal}
\keyword{transmit}
