% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrabid-package.R
\docType{package}
\name{simrabid-package}
\alias{simrabid}
\alias{simrabid-package}
\title{simrabid: Individual-based model of canine rabies}
\description{
This package implements a stochastic individual-based model of 
    canine rabies transmission. To learn more about simrabid, 
    start with the vignettes: browseVignettes(package = "simrabid").
}
\author{
\strong{Maintainer}: Malavika Rajeev \email{mrajeev@princeton.edu} (\href{https://orcid.org/0000-0001-6030-4995}{ORCID})

}
\keyword{internal}
