% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simrabid}
\alias{simrabid}
\title{Simulate a spatially-explicit, individual-based model of canine rabies}
\usage{
simrabid(
  start_up,
  start_vacc,
  I_seeds,
  vacc_dt,
  params = c(list(R0 = 1.2, k = 1, iota = 4), param_defaults),
  days_in_step = 7,
  observe_fun = beta_detect_monthly,
  serial_fun = serial_lognorm,
  dispersal_fun = steps_weibull,
  secondary_fun = nbinom_constrained,
  incursion_fun = sim_incursions_pois,
  movement_fun = sim_movement_continuous,
  sequential = TRUE,
  allow_invalid = TRUE,
  leave_bounds = TRUE,
  max_tries = 100,
  summary_fun = return_env,
  track = TRUE,
  weights = NULL,
  row_probs = NULL,
  coverage = FALSE,
  break_threshold = 0.8,
  by_admin = FALSE,
  extra_pars = NULL,
  routine_vacc = 0
)
}
\arguments{
\item{start_up}{objects for start up generated \code{\link{setup_sim}}}

\item{start_vacc}{numeric vector \link{0, 1}, starting proportion vaccinated
either length 1 or length of cells being tracked}

\item{I_seeds}{integer, initial number of infections to seed randomly}

\item{vacc_dt}{data.table with three columns: vacc_times (the timestep of vaccination,
should correspond to the same timesteps as being simulated),
vacc_ests (either coverage or number vaccinated), and
vacc_locs (the location id corresponding to the shapefile)}

\item{params}{parameter list, should include all parameters for simulation
functions (i.e. serial_, dispersal_, secondary_, incursion_, etc.)}

\item{days_in_step}{integer, the number of days in each time step being
simulated (i.e. 7 for weekly, 1 for daily)}

\item{observe_fun}{function for simulating observation/detection, should take
two parameters: I_dt (the line list of cases generated in simrabid) and
params (the list of params that will be passed through the \code{param})}

\item{serial_fun}{function for drawing the serial interval between cases}

\item{dispersal_fun}{function for drawing the dispersal distance/step length}

\item{secondary_fun}{function for generating secondary cases}

\item{incursion_fun}{function for seeding incursions}

\item{movement_fun}{function for simulating movement, current options are
\code{\link{sim_movement_continuous}} and \code{\link{sim_movement_prob}}}

\item{sequential}{boolean, whether movements should be sequential or
drawn as a kernel}

\item{allow_invalid}{boolean, whether movements to invalid locations
should be kept or movements should be resampled (i.e. then transmission
events can fail due to movement into an unihabited patch}

\item{leave_bounds}{boolean, whether movements to locations outside the bounds
of the area being simulated should be kept or those movements should be
resampled (i.e. transmission events can fail due to movement outside of
the district}

\item{max_tries}{the maximum number of times to resample invalid movements
if either or both \code{allow_invalid} and \code{leave_bounds} is false.}

\item{summary_fun}{function for summarizing simulation, should take two
parameters a character vector of object names from the simrabid environment
that you need for use in the function.}

\item{track}{boolean, whether to track failed transmission events in the line
list of cases (I_dt), i.e. invalid movements, contacts with V/E/I individuals}

\item{weights}{numeric vector of the length of the number of cells (both tracked
and untracked) in the raster being simulated (for use with
\code{\link{sim_movement_prob}}}

\item{row_probs}{numeric vector of the length of the number of tracked cells
for allocating vaccinations by (i.e. if you except vaccinations to be more
likely in certain grid cells vs. others)}

\item{coverage}{boolean, are the vaccination estimates number of
individuals (FALSE) or the proportion vaccinated (TRUE)}

\item{break_threshold}{numeric \link{0, 1}, whether to cut-off the simulation
if population declines to below this threshold (to stop run-away simulations)}

\item{by_admin}{boolean, whether simulation is being aggregated to the grid cell
or administrative (or arbitrary unit) based on the shapefile}

\item{extra_pars}{list, additional parameters that you might need for the summary
or any other custom functions}

\item{routine_vacc}{proportion of new born dogs vaccinated, defaults to zero}
}
\value{
output from function passed to \code{summary_fun}
}
\description{
This function simulates the rabies IBM at the specified scale. Inputs much
match between the \code{\link{setup_sim}} and \code{\link{setup_space}} functions
in terms of the space and timesteps being simulated.
}
