% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterize.R
\name{steps_weibull}
\alias{steps_weibull}
\title{Step length distribution of rabid animal movements}
\usage{
steps_weibull(n, params)
}
\arguments{
\item{n}{number to draw}

\item{params}{list of parameters}
}
\value{
numeric vector of step lengths of length \code{n}
}
\description{
This function draws step lengths (in meters)
from a lognormal distribution, pass \code{param_defaults} to the parameter list
for best-fit parameters from Mancy et al.
Can be passed to the \code{dispersal_fun} in \code{\link{simrabid}} (should
set parameter \code{sequential} to TRUE as well in that case).
}
