% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{use_mget}
\alias{use_mget}
\title{Return a list of named objects from parent environment}
\usage{
use_mget(names = c("I_dt", "I"), envir_num = 1)
}
\arguments{
\item{names}{character vector of object names in parent environment}
}
\value{
a list of objects from simrabid specified in names
}
\description{
This function is an example of a summary function that can be
passed to \code{simrabid}. If passing to the summary_fun arg in \code{simrabid},
it will be called with no arguments,
so all arguments must have a default or be empty. This is also a useful
helper function for writing your own summary functions, just wrap it in
list2env and set envir_num to the parent environment frame number where
the objects you're looking for are (see the source for \code{check_sim} for an example).
}
\keyword{summarize}
