% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce.R
\name{add_incursions}
\alias{add_incursions}
\title{Title}
\usage{
add_incursions(
  incs,
  cell_ids,
  x_coord,
  y_coord,
  tstep,
  counter,
  days_in_step = 7
)
}
\arguments{
\item{days_in_step}{}
}
\value{

}
\description{
Title
}
