% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_sim}
\alias{setup_sim}
\title{Set-up simulation}
\usage{
setup_sim(
  start_date = "2002-01-01",
  apprx_end_date = "2020-01-01",
  days_in_step = 7,
  rast,
  death_rate_annual,
  birth_rate_annual,
  waning_rate_annual,
  block_fun = block_cells,
  params = list(start_pop),
  by_admin = FALSE
)
}
\arguments{
\item{start_date}{date, the starting date for the simulation in year-month-day format}

\item{apprx_end_date}{date, the approximate end date for the simulation in year-month-day format,
the set-up fun will return the actual end date of the simulation rounded up to the
maximum number of timesteps in the interval based on \code{days_in_step}}

\item{days_in_step}{the number of days to aggregate timesteps by, defaults to 7 ("weekly"),
but can be arbitrary number of days}

\item{rast}{raster, from \code{\link{setup_space}} the output raster with each cell allocated
to a location}

\item{death_rate_annual}{annual death rate, either length 1 or length of the admin units}

\item{birth_rate_annual}{annual birth rate, either length 1 or length of the admin units}

\item{waning_rate_annual}{annual waning rate of vaccination immunity,
either length 1 or length of the admin units}

\item{block_fun}{a function that designates the indices to track/block/consider
out-of-bounds (see \code{\link{block_cells}} for an example)}

\item{params}{additional params to pass, must include start_pop, a numeric
vector of the starting population in each grid cell of \code{rast}}

\item{by_admin}{logical, should mixing be at the admin (or arbitrary) scale?}
}
\value{
a list of objects needed for the simulation
}
\description{
\code{setup_sim} takes inputs and generates all needed outputs to be passed
to \code{\link{simrabid}}.
}
\keyword{setup}
