% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmit.R
\name{sim_bites}
\alias{sim_bites}
\title{Simulates biting/exposure behavior of currently infected individuals}
\usage{
sim_bites(
  secondaries,
  ids = I_now$id,
  x_coords = I_now$x_coord,
  y_coords = I_now$y_coord,
  t_infectious = I_now$t_infectious,
  counter = max(I_dt$id),
  dispersal_fun,
  res_m,
  row_ids,
  cell_ids,
  cells_pop,
  nrow,
  ncol,
  x_topl,
  y_topl,
  sequential = TRUE,
  allow_empties = TRUE,
  leave_bounds = TRUE,
  max_tries = 100
)
}
\arguments{
\item{secondaries}{numeric vector of number of secondary cases for each infectious individual}

\item{ids}{numeric vector of the ids of infected individuals (progenitor ids)}

\item{x_coords}{numeric vector, x coordinate (Easting) of infected individuals}

\item{y_coords}{numeric vector, y coordinate (Northing) of infected individuals}

\item{t_infectious}{numeric vector, the time (fractional time step) at which each
individual became infectious}

\item{counter}{integer, the id number to start with for resulting secondary cases}

\item{dispersal_fun}{function with a single parameter, n, which will draw distances
for each individual to move}

\item{res_m}{numeric, the grid cell resolution in meters}

\item{row_ids}{integer vector, the row ids corresponding to each grid cell in which
the infectious individual is starting from}

\item{cell_ids}{integer vector, the cell ids corresponding to each grid cell in which
the infectious individual is starting from}

\item{cells_pop}{integer vector, the cell ids of grid cells that are populated
or generally where movements are valid to}

\item{nrow}{numeric, the number of rows in the grid}

\item{ncol}{numeric, the number of columns in the grid}

\item{x_topl}{numeric, the top left x coordinate of the grid on which
movement is being simulated}

\item{y_topl}{numeric, the top left y coordinate of the grid on which
movement is being simulated}

\item{sequential}{boolean, if TRUE then movements are sequential, if FALSE, then
movements are kernel based (see Details)}

\item{allow_empties}{boolean, are movements to empty patches (i.e. with no dogs) valid}

\item{leave_bounds}{boolean, are movements to outside of the boundaries of the are being
simulated valid}

\item{max_tries}{integer, the maximum number of tries to make before accepting
an invalid movement (i.e. transmission event fails due to either leaving the
bounds of the simulation or moving to an uninhabitable grid cell) if either
or both leave_bounds and allow_empties are FALSE.}
}
\value{
a data.table that corresponds to the columns in I_dt in the simulation.
See \code{\link{simrabid}} for full description.
}
\description{
\code{sim_bites} simulates spatially-explicit transmission events per
infectious individual.
}
\details{
This function takes the currently infectious individuals and the draws of the number of
secondary cases that they seed and simulates their movement.
Movements can be sequential (i.e. infected individuals moves from origin to
first location of exposure, then from this to the second, etc.)
or can be kernel based where movements are all seeded from
the origin location of the infected individual. Movements outside of the bounds
of the simulation or to uninhabitable grid cells can either be accepted (if
\code{leave_bounds} or \code{allow_empties} are \code{TRUE}). If \code{FALSE}, these movements will not be
accepted and will be redrawn \code{max_tries} times.
}
\keyword{internal}
\keyword{transmit}
