% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduce.R
\name{sim_incursions_pois}
\alias{sim_incursions_pois}
\title{Simulate incursions}
\usage{
sim_incursions_pois(cell_ids, params = list(iota = 1))
}
\arguments{
\item{cell_ids}{the cells to sample incursions to}

\item{params}{a list with a named parameter \code{iota}, the number of incursions
on average in each time step}
}
\value{
numeric vector of cell_ids where incursions occured
}
\description{
\code{sim_incursions_pois} simulates incursions from a poisson distribution.
}
\details{
These functions are passed as an argument to the parameter \code{incursion_fun} in
\code{\link{simrabid}} and can be customized. They must include arguments
for \code{cell_ids} and \code{params} (if you want to change the parameter
values or you can fix them within the function).
}
\keyword{incursions}
